refactored menu code
